// src/types.ts
export enum Tool {
  DASHBOARD = 'DASHBOARD',
  DUBBING_STUDIO = 'DUBBING_STUDIO',
  AUDIO_ENHANCER = 'AUDIO_ENHANCER',
  PODCAST_EDITOR = 'PODCAST_EDITOR',
  SUBTITLE_GENERATOR = 'SUBTITLE_GENERATOR',
  AI_MUSIC_COMPOSER = 'AI_MUSIC_COMPOSER',
  SCRIPT_DOCTOR = 'SCRIPT_DOCTOR',
  TEXT_TO_SPEECH = 'TEXT_TO_SPEECH',
  STEM_SPLITTER = 'STEM_SPLITTER',
  SFX_GENERATOR = 'SFX_GENERATOR',
}